Update infrastructure server backups
* Enhance mediawiki backup script to do better error handling
* Add Git repository backup for Gerrit

+++ modules/traf/files/mediawiki/backupWiki.sh
@@ -25,28 +25,35 @@ cd $wikiDir
 
 # create temporary tarball
 tar cfz $backupDir/tmp.wiki.tgz w
-md5sum $backupDir/tmp.wiki.tgz > $backupDir/tmp.wiki.md5
-diff -q $backupDir/wiki.md5 $backupDir/tmp.wiki.md5 2>/dev/null
-
-# if there is a difference in the md5sum of the tarball
-# archive the old one if it exists, rename temporary files
-# then upload tar file to object storage
-if [ $? -ne 0 ]
+tarRes=$?
+fSize=$(stat -c%s "$backupDir/tmp.wiki.tgz")
+if [ $tarRes -eq 0 -a $fSize -ne 0 ]
 then
-  if [ -f $backupDir/wiki.tgz ]
-  then
-    cur_date=$(date +%Y%m%d)
-    mv $backupDir/wiki.md5 $backupDir/wiki.md5.$cur_date
-    mv $backupDir/wiki.tgz $backupDir/wiki.tgz.$cur_date
-  fi
+    md5sum $backupDir/tmp.wiki.tgz > $backupDir/tmp.wiki.md5
+    diff -q $backupDir/wiki.md5 $backupDir/tmp.wiki.md5 2>/dev/null
 
-  mv $backupDir/tmp.wiki.md5 $backupDir/wiki.md5
-  mv $backupDir/tmp.wiki.tgz $backupDir/wiki.tgz
-  backupToObjectStorage.sh upload $backupDir/wiki.tgz
+    # if there is a difference in the md5sum of the tarball
+    # archive the old one if it exists, rename temporary files
+    # then upload tar file to object storage
+    if [ $? -ne 0 ]
+    then
+      if [ -f $backupDir/wiki.tgz ]
+      then
+        cur_date=$(date +%Y%m%d)
+        mv $backupDir/wiki.md5 $backupDir/wiki.md5.$cur_date
+        mv $backupDir/wiki.tgz $backupDir/wiki.tgz.$cur_date
+      fi
 
-# if there is no difference in the md5sum of the tarball
-# then remove all temporary files
+      mv $backupDir/tmp.wiki.md5 $backupDir/wiki.md5
+      mv $backupDir/tmp.wiki.tgz $backupDir/wiki.tgz
+      backupToObjectStorage.sh upload $backupDir/wiki.tgz
+
+    # if there is no difference in the md5sum of the tarball
+    # then remove all temporary files
+    else
+      rm $backupDir/tmp.wiki.tgz
+      rm $backupDir/tmp.wiki.md5
+    fi
 else
-  rm $backupDir/tmp.wiki.tgz
-  rm $backupDir/tmp.wiki.md5
+    echo "ERROR: $0 tar encountered an error or generated a 0 size tarball. Please check the backup job."
 fi
